pyproject.toml
[tool.poetry]
name = "deeplabv3-demo"
version = "0.2.0"
description = ""
authors = ["None"]
license = "MIT"
readme = "README.md"
packages = [{include = "deeplabv3_demo"}]

[tool.poetry.dependencies]
python = ">=3.8,<3.10"
argparse = "^1.4.0"
numpy = "^1.24.2"
pillow = "^9.4.0"
opencv-python = "^4.7.0.68"
tflite-runtime = "^2.11.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.2.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
demo = "deeplabv3_demo.demo:demo_entrypoint"

[tool.black]
target-version = ['py39']
line-length = 120
include_trailing_comma = false

# PyLint
[tool.pylint.MAIN]

# Specify a score threshold under which the program will exit with error.
fail-under=10

# Use multiple processes to speed up Pylint. Specifying 0 will auto-detect the
# number of processors available to use, and will cap the count on Windows to
# avoid hangs.
jobs=0

# Files or directories to be skipped. They should be base names, not paths.
ignore=["dist", "resources", "tests"]

# Discover python modules and packages in the file system subtree.
recursive=true

[tool.pylint.TYPECHECK]

# List of members which are set dynamically and missed by pylint inference
# system, and so shouldn't trigger E1101 when accessed. Python regular
# expressions are accepted.
generated-members= [
    "cv2.*"
]

[tool.pylint.'MESSAGES CONTROL']
# Disable the message, report, category or checker with the given id(s). You
# can either give multiple identifier separated by comma (,) or put this option
# multiple time (only on the command line, not in the configuration file where
# it should appear only once).
disable=["C0116", "C0114", "C0413", "E1101", "R0914", "W0702", "W1201", "R0903", "E0401", "W0401"]

# pyproject.toml
[tool.pytest.ini_options]
minversion = "7.0"
addopts = "-ra -q -p no:warnings --verbose -color=yes"
testpaths = [
    "tests",
]
